Adjust Justfile just so it is just so
This commit reorders the Justfile recipes so that they appear in a
topologically sorted order. That is, if the recipes are trees of
dependencies, then the leaves must appear first, and any non-leaf recipe
must appear before the non-leaf recipes that use it. This helps the
developer to test the recipes in an order that is efficient, and helps
the user understand which recipes are likely used by which other recipes
(earlier ones are used by later ones, only).

This commit also removes the 'build' target, which, interestingly, is
not used by any other recipes or in production.

The unnecessary 'build' is also removed from the 'test' target,
hopefully making it faster.